Readme changes and SCRM now in separate file
